requirements.txt
openai==1.50.2
requests==2.32.3
streamlit==1.38.0
moviepy==1.0.3
pydub==0.25.1
ffmpeg-python==0.2.0
python-dotenv==1.0.1
pytest==8.3.2
pytest-mock==3.14.0
pydantic==2.9.2
tenacity==8.5.0
srt==3.5.3
Pillow==10.4.0
pyttsx3==2.90
